Move TriggeredBuffer to its own file
